Rename `util` package to `yangentry` package.
Per https://go.dev/blog/package-names we should try to avoid using a
generic package name.

Since `Entry` is the primary output users are interested in I thought
this is not too bad of a name.

Fixes #213
