+++ pkg/yangentry/build_yang_test.go
@@ -12,14 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package util
+package yangentry
 
 import "testing"
 
-// TestProcessModules tests the ProcessModules function - which takes an input
+// TestParse tests the Parse function - which takes an input
 // set of modules and processes them using the goyang compiler into a set of
 // yang.Entry pointers.
-func TestProcessModules(t *testing.T) {
+func TestParse(t *testing.T) {
 	tests := []struct {
 		name     string
 		inFiles  []string
@@ -59,7 +59,7 @@ func TestProcessModules(t *testing.T) {
 	}}
 
 	for _, tt := range tests {
-		entries, errs := ProcessModules(tt.inFiles, tt.inPath)
+		entries, errs := Parse(tt.inFiles, tt.inPath)
 		if len(errs) != 0 && !tt.wantErr {
 			t.Errorf("%s: unexpected error processing modules: %v", tt.name, errs)
 			continue
