+++ pkg/yangentry/testdata/00-valid-module.yang
