+++ pkg/yangentry/testdata/04-valid-module-one.yang
