+++ pkg/yangentry/testdata/04-valid-module-two.yang
